SOF-335: Split SettingsTile into SettingsInfoTile and SettingsActionTile
